requirements.txt
fastapi[all]
python-jose
passlib
bcrypt
pymongo
python-multipart